Splitting classes into packages
