--- OnlineStore/consoleApp/src/main/java/Commands.java
@@ -1,7 +1,0 @@
-public class Commands {
-    public static void quit(String string){
-        if (string.equals("quit")){
-            System.exit(0);
-        }
-    }
-}
