+++ OnlineStore/consoleApp/src/main/java/com/marina/Commands.java
@@ -1,0 +1,9 @@
+package com.marina;
+
+public class Commands {
+    public static void quit(String string){
+        if (string.equals("quit")){
+            System.exit(0);
+        }
+    }
+}
