+++ OnlineStore/domain/src/main/java/com/marina/dao/builder/QueryBuilder.java
@@ -1,0 +1,16 @@
+package com.marina.dao.builder;
+
+public class QueryBuilder {
+    public static String getAllCategories(){
+        return "SELECT * FROM category";
+    }
+
+    public static  String deleteAllCategories(){
+        return "DELETE FROM category";
+    }
+
+    public static String insertProductQuery(){
+        return "INSERT INTO product (name, rating, price, categoryID, quantity) VALUES(" +
+                "?, ?, ?, ?, ?)";
+    }
+}
