+++ OnlineStore/domain/src/main/java/com/marina/dao/exceptions/DataException.java
@@ -1,0 +1,10 @@
+package com.marina.dao.exceptions;
+
+public class DataException extends RuntimeException {
+    public DataException(String msg){
+        super(msg);
+    }
+    DataException(String msg, Exception e){
+        super(msg, e);
+    }
+}
