+++ OnlineStore/store/src/main/java/com/marina/ChainComparator.java
@@ -1,0 +1,25 @@
+package com.marina;
+
+import com.marina.dao.domain.Product;
+import java.util.Comparator;
+import java.util.List;
+
+public class ChainComparator implements Comparator<Product> {
+    private List<Comparator<Product>> comparatorList;
+
+    public ChainComparator(List<Comparator<Product>> comparatorList){
+        this.comparatorList = comparatorList;
+    }
+
+    @Override
+    public int compare (Product p1, Product p2) {
+        int result;
+        for (Comparator<Product> comparator : comparatorList) {
+            if ((result = comparator.compare(p1, p2)) != 0) {
+                return result;
+            }
+        }
+        return  0;
+    }
+
+}
