+++ OnlineStore/store/src/main/java/com/marina/domain/Sort.java
@@ -1,0 +1,35 @@
+package com.marina.domain;
+
+public class Sort {
+    private String name, price, rating;
+    public Sort(){};
+    public Sort(String name, String price, String rating){
+        this.name = name;
+        this.price = price;
+        this.rating = rating;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getPrice() {
+        return price;
+    }
+
+    public void setPrice(String price) {
+        this.price = price;
+    }
+
+    public String getRating() {
+        return rating;
+    }
+
+    public void setRating(String rating) {
+        this.rating = rating;
+    }
+}
