--- OnlineStore/store/src/main/java/exceptions/DataException.java
@@ -1,10 +1,0 @@
-package exceptions;
-
-public class DataException extends RuntimeException {
-    public DataException(String msg){
-        super(msg);
-    }
-    DataException(String msg, Exception e){
-        super(msg, e);
-    }
-}
