--- OnlineStore/store/src/main/java/other/Sort.java
@@ -1,35 +1,0 @@
-package other;
-
-public class Sort {
-    private String name, price, rating;
-    public Sort(){};
-    public Sort(String name, String price, String rating){
-        this.name = name;
-        this.price = price;
-        this.rating = rating;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getPrice() {
-        return price;
-    }
-
-    public void setPrice(String price) {
-        this.price = price;
-    }
-
-    public String getRating() {
-        return rating;
-    }
-
-    public void setRating(String rating) {
-        this.rating = rating;
-    }
-}
