--- OnlineStore/store/src/main/resources/db.properties
@@ -1,6 +1,0 @@
-HOST=127.0.0.1
-PORT=3306
-URL=jdbc:mysql://localhost:3306/store?useUnicode=true&serverTimezone=UTC
-USER=root
-PWD=11111Aa!
-DBNAME=store
